rl_agent/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
aiokafka==0.8.1